Fix OpenIDC shim for HTTP headers too large error
Auth0 sends information back to the OpenIDC shim which gets converted to
HTTP headers and sent to Jenkins. A recent change increased the amount
of data sent, which in turn caused the amount of HTTP headers to exceed
the hardcoded limit in Jenkins. So we stop using the default shim
offered by EIS and use a local copy that trims down the headers.